requirements.txt
openai
azure-search-documents
azure-core
Flask==3.0.3
requests